Bug fix: opendna.py printing predicted 2d string
Line268: change the key in the self.ssAnalysis which is used to output the 2d string of the predicted 2d structure.
Line357: in nupack, add 'displayed 2d string' to self.ssAnalysis dictionary to output the dot-parenthesis notation of the top 2d structure. Previously, self.Analysis['2d string] is used but the order of 2d strings in it is wrong.